Added a source_type and message field to ValidationError.
- ValidationError should have more detail.
- ValidationError::message should maybe have a String so people can use
format! to generate the messages.
