+++ prae/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "prae"
-version = "0.4.4"
+version = "0.4.5"
 authors = ["Alex Ryapolov <ryapolov@pm.me>"]
 edition = "2018"
 license = "Unlicense"
@@ -12,10 +12,10 @@ keywords = ["validation", "macro", "invariant", "generation"]
 categories = ["development-tools"]
 
 [dependencies]
-prae_macro = { version = "0.2.1", path = "../prae_macro" }
-serde = { version = "1.0.126", optional = true }
-thiserror = "1.0.25"
+prae_macro = { version = "0.2", path = "../prae_macro" }
+serde = { version = "1.0", optional = true }
 
 [dev-dependencies]
-serde_json = "1.0.64"
-serde = { version = "1.0.126", features = ["derive"] }
+assert_matches = "1.5"
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
